MongoDB and Zen Services Backup

+++ other-scripts/bar/020-barFullBackup.sh
@@ -88,7 +88,7 @@ echo
 
 
 ##### Preparation ##############################################################
-# Verify OCP Connecction
+## Verify OCP Connection
 logInfo "Verifying OC CLI is connected to the OCP cluster..."
 WHOAMI=$(oc whoami)
 logInfo "WHOAMI =" $WHOAMI
@@ -100,7 +100,7 @@ if [[ "$WHOAMI" == "" ]]; then
 fi
 echo
 
-# switch to CP4BA project
+## Switch to CP4BA project
 project=$(oc project --short)
 logInfo "project =" $project
 if [[ "$project" != "$cp4baProjectName" ]]; then
@@ -109,16 +109,16 @@ if [[ "$project" != "$cp4baProjectName" ]]; then
 fi
 echo
 
-# Create backup directory
+## Create backup directory
 logInfo "Creating backup directory..."
 BACKUP_DIR=$BACKUP_ROOT_DIRECTORY_FULL/backup_${DATETIMESTR}
 mkdir -p $BACKUP_DIR
 
-# Get CP4BA depoyment name
+## Get CP4BA deployment name
 CP4BA_NAME=$(oc get ICP4ACluster -o name |cut -d "/" -f 2)
 logInfo "CP4BA deployment name: $CP4BA_NAME"
 
-# Get CP4BA version
+## Get CP4BA version
 CP4BA_VERSION=$(oc get ICP4ACluster $CP4BA_NAME -o 'custom-columns=NAME:.metadata.name,VERSION:.spec.appVersion' --no-headers | awk '{print $2}')
 logInfo "Found CP4BA version: $CP4BA_VERSION"
 echo
@@ -127,23 +127,128 @@ echo
 NAMESPACE_UID=$(oc describe project $cp4baProjectName | grep uid-range | cut -d"=" -f2 | cut -d"/" -f1)
 logInfo "Namespace $cp4baProjectName uid: $NAMESPACE_UID"
 echo $NAMESPACE_UID > ${BACKUP_DIR}/namespace_uid
+##### CPfs #####################################################################
+
+if [[ $CP4BA_VERSION =~ "21.0.3" ]]; then
+  ## Take MongoDB Backup 
+  # Prime templates 
+  logInfo "Priming resources to take MongoDB backup"
+  cp ${CUR_DIR}/templates/mongodb-backup-pvc.template.yaml ${CUR_DIR}/mongodb-backup-pvc.yaml
+  cp ${CUR_DIR}/templates/mongodb-backup-deployment.template.yaml ${CUR_DIR}/mongodb-backup-deployment.yaml
+
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g; s|§pvcStorageClass|$pvcStorageClassName|g" ${CUR_DIR}/mongodb-backup-pvc.yaml
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g" ${CUR_DIR}/mongodb-backup-deployment.yaml
+
+  # Create resources necessary to backup MongoDB 
+  oc apply -f ${CUR_DIR}/mongodb-backup-pvc.yaml
+  oc apply -f ${CUR_DIR}/mongodb-backup-deployment.yaml
+
+  # Wait indefinitely for deployment to be Available (pod Ready)
+  oc wait -f ${CUR_DIR}/mongodb-backup-deployment.yaml --for=condition=Available --timeout=-1s
+
+  mongodbpods=$(oc get pod -l=foundationservices.cloudpak.ibm.com=mongo-data --no-headers --ignore-not-found | awk '{print $1}' | sort)
+
+  for pod in ${mongodbpods[*]}
+  do
+    logInfo "Backing up MongoDB in from pod ${pod}..."
+    # prep certs files that will be used to take the backup
+    oc exec $pod -it -- bash -c 'cat /cred/mongo-certs/tls.crt /cred/mongo-certs/tls.key > /work-dir/mongo.pem; cat /cred/cluster-ca/tls.crt /cred/cluster-ca/tls.key > /work-dir/ca.pem'
+    oc exec $pod -it -- bash -c 'mongodump --oplog --archive=/dump/mongo.archive --host mongodb:$MONGODB_SERVICE_PORT --username $ADMIN_USER --password $ADMIN_PASSWORD --authenticationDatabase admin --ssl --sslCAFile /work-dir/ca.pem --sslPEMKeyFile /work-dir/mongo.pem'
+    oc cp $pod:/dump/mongo.archive ${BACKUP_DIR}/mongodb/mongo.archive
+    break
+  done
+
+  # check if backup was taken successfully
+  if [ -e "${BACKUP_DIR}/mongodb/mongo.archive" ]; then
+    # clean up pod storage and notify successful completion
+    logInfo "MongoDB backup completed successfully."
+    oc delete -f ${CUR_DIR}/mongodb-backup-deployment.yaml
+    oc delete -f ${CUR_DIR}/mongodb-backup-pvc.yaml
+  else
+    logError "MongoDB backup failed, check logs!"
+    exit 1
+  fi
+
+  ## Take Zen Services Backup
+  # Prime templates 
+  logInfo "Priming resources to take Zen Services backup"
+  cp ${CUR_DIR}/templates/zen-backup-pvc.template.yaml ${CUR_DIR}/zen-backup-pvc.yaml
+  cp ${CUR_DIR}/templates/zen4-br-scripts.template.yaml ${CUR_DIR}/zen4-br-scripts.yaml
+  cp ${CUR_DIR}/templates/zen4-sa.template.yaml ${CUR_DIR}/zen4-sa.yaml
+  cp ${CUR_DIR}/templates/zen4-role.template.yaml ${CUR_DIR}/zen4-role.yaml
+  cp ${CUR_DIR}/templates/zen4-rolebinding.template.yaml ${CUR_DIR}/zen4-rolebinding.yaml
+  cp ${CUR_DIR}/templates/zen-backup-deployment.template.yaml ${CUR_DIR}/zen-backup-deployment.yaml
+
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g; s|§pvcStorageClass|$pvcStorageClassName|g" ${CUR_DIR}/zen-backup-pvc.yaml
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g" ${CUR_DIR}/zen4-br-scripts.yaml
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g" ${CUR_DIR}/zen4-sa.yaml
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g" ${CUR_DIR}/zen4-role.yaml
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g" ${CUR_DIR}/zen4-rolebinding.yaml
+  sed -i.bak "s|§cp4baProjectNamespace|$cp4baProjectName|g" ${CUR_DIR}/zen-backup-deployment.yaml
+
+  # Create resources necessary to backup Zen Services
+  oc apply -f ${CUR_DIR}/zen-backup-pvc.yaml
+  oc apply -f ${CUR_DIR}/zen4-br-scripts.yaml
+  oc apply -f ${CUR_DIR}/zen4-sa.yaml
+  oc apply -f ${CUR_DIR}/zen4-role.yaml
+  oc apply -f ${CUR_DIR}/zen4-rolebinding.yaml
+  oc apply -f ${CUR_DIR}/zen-backup-deployment.yaml
+
+  # Wait indefinitely for deployment to be Available (pod Ready)
+  oc wait -f ${CUR_DIR}/zen-backup-deployment.yaml --for=condition=Available --timeout=-1s
+
+  zenbkpods=$(oc get pod -l=foundationservices.cloudpak.ibm.com=zen-data --no-headers --ignore-not-found | awk '{print $1}')
+
+  for pod in ${zenbkpods[*]}
+  do
+    logInfo "Backing up Zen Services from pod ${pod}..."
+    oc exec $pod -it -- bash -c "/zen4/zen4-br.sh $cp4baProjectName true"
+    # Go into directory where backup was created and compress into single file
+    oc exec $pod -it -- bash -c "cd /user-home && tar -cvf zen-metastoredb-backup.tar zen-metastoredb-backup"
+    oc cp $pod:/user-home/zen-metastoredb-backup.tar ${BACKUP_DIR}/zenbackup/zen-metastoredb-backup.tar
+    break
+  done
+
+  # check if backup was taken successfully
+  if [ -e "${BACKUP_DIR}/zenbackup/zen-metastoredb-backup.tar" ]; then
+    logInfo "Zen Services backup completed successfully."
+
+    oc delete -f ${CUR_DIR}/zen-backup-deployment.yaml
+    oc delete -f ${CUR_DIR}/zen4-rolebinding.yaml
+    oc delete -f ${CUR_DIR}/zen4-role.yaml
+    oc delete -f ${CUR_DIR}/zen4-sa.yaml
+    oc delete -f ${CUR_DIR}/zen4-br-scripts.yaml
+    oc delete -f ${CUR_DIR}/zen-backup-pvc.yaml
+  else
+    logError "Zen Services backup failed, check logs!"
+    exit 1
+  fi
+else
+  # Backup implementation of CPfs for CP4BA versions 22, 23, and 24 is not developed yet. 
+  # TODO: Backup CPfs for other versions of CP4BA specially 24
+  logError "Do not know how to take backup of CPfs services for this Cloud Pak version $CP4BA_VERSION"
+  exit 1
+fi
 
 ##### Backup BTS PostgreSQL Database ###########################################
-btscnpgpods=$(oc get pod -l=app.kubernetes.io/name=ibm-bts-cp4ba-bts --no-headers --ignore-not-found | awk '{print $1}')
+btscnpgpods=$(oc get pod -l=app.kubernetes.io/name=ibm-bts-cp4ba-bts --no-headers --ignore-not-found | awk '{print $1}' | sort)
 for pod in ${btscnpgpods[*]}
 do
-  if [[ "$pod" = "ibm-bts-cnpg-"$cp4baProjectName"-cp4ba-bts-1" ]]; then
-    logInfo "Backing up BTS PostgreSQL Database..."
-    oc exec --container postgres ibm-bts-cnpg-${cp4baProjectName}-cp4ba-bts-1 -it -- bash -c "pg_dump -d BTSDB -U postgres -Fp -c -C --if-exists  -f /var/lib/postgresql/data/backup_btsdb.sql"
-    oc cp --container postgres ibm-bts-cnpg-${cp4baProjectName}-cp4ba-bts-1:/var/lib/postgresql/data/backup_btsdb.sql ${BACKUP_DIR}/postgresql/backup_btsdb.sql
-    oc exec --container postgres ibm-bts-cnpg-${cp4baProjectName}-cp4ba-bts-1 -it -- bash -c "rm -f /var/lib/postgresql/data/backup_btsdb.sql"
-    
-    # After the backup, we also can delete this pod
-    logInfo "Scaling down last BTS PostgreSQL Database pod..."
-    logInfo $(oc delete pod "ibm-bts-cnpg-"$cp4baProjectName"-cp4ba-bts-1")
-    echo
-  fi
+  logInfo "Backing up BTS PostgreSQL Database from pod ${pod}..."
+  oc exec --container postgres $pod -it -- bash -c "pg_dump -d BTSDB -U postgres -Fp -c -C --if-exists  -f /var/lib/postgresql/data/backup_btsdb.sql"
+  oc cp --container postgres ${pod}:/var/lib/postgresql/data/backup_btsdb.sql ${BACKUP_DIR}/postgresql/backup_btsdb.sql
+  break
 done
+
+# check if backup was taken successfully
+if [ -e "${BACKUP_DIR}/postgresql/backup_btsdb.sql" ]; then
+  logInfo "BTS PostgreSQL Database backup completed successfully."
+  # clean up pod storage
+  oc exec --container postgres $pod -it -- bash -c "rm -f /var/lib/postgresql/data/backup_btsdb.sql"
+else
+  logError "BTS PostgreSQL Database backup failed, check logs!"
+  exit 1
+fi
 
 ##### BAI ######################################################################
 # Take ES snapshot
@@ -202,48 +307,10 @@ if [[ "$MANAGEMENT_POD" != "" ]]; then
     logInfo "  Copying the savepoint to ${BACKUP_DIR}/flink/${FLINK_SAVEPOINT_LOCATION}..."
     oc cp --container management ${MANAGEMENT_POD}:${FLINK_SAVEPOINT_LOCATION} ${BACKUP_DIR}/flink${FLINK_SAVEPOINT_LOCATION}
   done
-  
-  # Remove the Flink job submitters
-  logInfo "Removing flink job submitters..."
-  oc get jobs -o custom-columns=NAME:.metadata.name | grep bai- | grep -v bai-setup | xargs oc delete job
-  echo
-  
-  # After the backup, we also can delete these pods
-  logInfo "Scaling down insights-engine pods..."
-  logInfo $(oc scale deployment.apps/iaf-insights-engine-management --replicas=0);
-  sleep 5
-  sts=$(oc get statefulset.apps/iaf-system-kafka -o name --ignore-not-found)
-  if [[ "$sts" = "" ]]; then
-    logInfo $(oc delete pod iaf-system-kafka-0)
-  else
-    logInfo $(oc scale statefulset.apps/iaf-system-kafka --replicas=0);
-  fi
-  sleep 5
-  sts=$(oc get statefulset.apps/iaf-system-zookeeper -o name --ignore-not-found)
-  if [[ "$sts" = "" ]]; then
-    logInfo $(oc delete pod iaf-system-zookeeper-0)
-  else
-    logInfo $(oc scale statefulset.apps/iaf-system-zookeeper --replicas=0);
-  fi
-  echo
-  sleep 5
 fi
 
 # Create ES/OS snapshots
-if [[ $CP4BA_VERSION =~ "24.0" ]]; then
-  # OpenSearch for 24.0 and later
-  logInfo "Declaring the location of the snapshot repository for OpenSearch..."
-  OPENSEARCH_ROUTE=$(oc get route opensearch-route -o jsonpath='{.spec.host}')
-  OPENSEARCH_PASSWORD=$(oc get secret opensearch-ibm-elasticsearch-cred-secret --no-headers --ignore-not-found -o jsonpath={.data.elastic} | base64 -d)
-  curl -skl -u elastic:$OPENSEARCH_PASSWORD -XPUT "https://$OPENSEARCH_ROUTE/_snapshot/${DATETIMESTR}" -H "Content-Type: application/json" -d'{"type":"fs","settings":{"location": "/workdir/snapshot_storage","compress": true}}'
-  logInfo "Creating snapshot backup_${DATETIMESTR}..."
-  SNAPSHOT_RESULT=$(curl -skL -u elastic:${OPENSEARCH_PASSWORD} -XPUT "https://${OPENSEARCH_ROUTE}/_snapshot/${DATETIMESTR}/backup_${DATETIMESTR}?wait_for_completion=true&pretty=true")
-  SNAPSHOT_STATE=$(echo $SNAPSHOT_RESULT | jq -r ".snapshot.state")
-  checkResult $SNAPSHOT_STATE "SUCCESS" "Snapshot state"
-  echo
-  # TODO: copy the snapshots from the pod, what should be copied ? Need clarification from document, Zhong Tao opened a case for this issue: https://jsw.ibm.com/browse/DBACLD-164204: Need clarification on how to copy ES/OS snapshots to another environment
-  # TODO: scale down es pods
-else
+if [[ $CP4BA_VERSION =~ "21.0.3" ]]; then
   # ElasticSearch
   logInfo "Declaring the location of the snapshot repository for ElasticSearch..."
   ELASTICSEARCH_ROUTE=$(oc get route iaf-system-es -o jsonpath='{.spec.host}')
@@ -261,35 +328,34 @@ else
   # TODO: Next commands produce some console output that we might want to get rid of
   oc exec --container elasticsearch iaf-system-elasticsearch-es-data-0 -it -- bash -c "tar -cvf /usr/share/elasticsearch/es_snapshots_main_backup_${DATETIMESTR}.tgz /usr/share/elasticsearch/snapshots/main"
   oc cp --container elasticsearch iaf-system-elasticsearch-es-data-0:/usr/share/elasticsearch/es_snapshots_main_backup_${DATETIMESTR}.tgz ${BACKUP_DIR}/es_snapshots_main_backup_${DATETIMESTR}.tgz
-  # Delete the tar, the snapshot and the repository
-  oc exec --container elasticsearch iaf-system-elasticsearch-es-data-0 -it -- bash -c "rm -f /usr/share/elasticsearch/es_snapshots_main_backup_${DATETIMESTR}.tgz"
-  curl -skL -u elasticsearch-admin:${ELASTICSEARCH_PASSWORD} -X DELETE "https://$ELASTICSEARCH_ROUTE/_snapshot/${DATETIMESTR}/backup_${DATETIMESTR}?pretty"
-  curl -skL -u elasticsearch-admin:${ELASTICSEARCH_PASSWORD} -X DELETE "https://$ELASTICSEARCH_ROUTE/_snapshot/${DATETIMESTR}?pretty"
-  
-  # After the backup, we also can delete these pods
-  logInfo "Scaling down es pods..."
-  logInfo $(oc scale statefulset.apps/iaf-system-elasticsearch-es-data --replicas=0);
+
+  # check if backup was taken successfully
+  if [ -e "${BACKUP_DIR}/es_snapshots_main_backup_${DATETIMESTR}.tgz" ]; then
+    logInfo "Elasticsearch backup completed successfully."
+    # Clean up, delete the tar, the snapshot and the repository
+    oc exec --container elasticsearch iaf-system-elasticsearch-es-data-0 -it -- bash -c "rm -f /usr/share/elasticsearch/es_snapshots_main_backup_${DATETIMESTR}.tgz"
+    curl -skL -u elasticsearch-admin:${ELASTICSEARCH_PASSWORD} -X DELETE "https://$ELASTICSEARCH_ROUTE/_snapshot/${DATETIMESTR}/backup_${DATETIMESTR}?pretty"
+    curl -skL -u elasticsearch-admin:${ELASTICSEARCH_PASSWORD} -X DELETE "https://$ELASTICSEARCH_ROUTE/_snapshot/${DATETIMESTR}?pretty"
+  else
+    logError "Elasticsearch backup failed, check the logs!"
+    exit 1
+  fi
+else # 24.0 and greater
+  # OpenSearch for 24.0 and later
+  logInfo "Declaring the location of the snapshot repository for OpenSearch..."
+  OPENSEARCH_ROUTE=$(oc get route opensearch-route -o jsonpath='{.spec.host}')
+  OPENSEARCH_PASSWORD=$(oc get secret opensearch-ibm-elasticsearch-cred-secret --no-headers --ignore-not-found -o jsonpath={.data.elastic} | base64 -d)
+  curl -skl -u elastic:$OPENSEARCH_PASSWORD -XPUT "https://$OPENSEARCH_ROUTE/_snapshot/${DATETIMESTR}" -H "Content-Type: application/json" -d'{"type":"fs","settings":{"location": "/workdir/snapshot_storage","compress": true}}'
+  logInfo "Creating snapshot backup_${DATETIMESTR}..."
+  SNAPSHOT_RESULT=$(curl -skL -u elastic:${OPENSEARCH_PASSWORD} -XPUT "https://${OPENSEARCH_ROUTE}/_snapshot/${DATETIMESTR}/backup_${DATETIMESTR}?wait_for_completion=true&pretty=true")
+  SNAPSHOT_STATE=$(echo $SNAPSHOT_RESULT | jq -r ".snapshot.state")
+  checkResult $SNAPSHOT_STATE "SUCCESS" "Snapshot state"
   echo
-  sleep 5
+  # TODO: copy the snapshots from the pod, what should be copied ? Need clarification from document, Zhong Tao opened a case for this issue: https://jsw.ibm.com/browse/DBACLD-164204: Need clarification on how to copy ES/OS snapshots to another environment
+  # TODO: scale down os pods
 fi
 
-# TODO: We have to check if this could stay as is...we maybe want to wait here, till admin has removed the other pods, otherwise the backup will be incomplete / distributed onto multiple backups
-# Wait till all pods are gone
-for ((i=0; i<13; i++)); do
-   remainingPods=$(oc get pods --no-headers -o name)
-   if [[ $(oc get pods --no-headers -o name) = "" ]]; then
-     break
-   else
-     if [[ $i = 12 ]]; then
-       logError "Remaining pods detected. Please check. Exiting WITHOUT full backup!!!"
-       echo
-       exit 1
-     fi
-     sleep 10
-   fi
-done
-
-# Now that those backups are done, and all remaining pods are gone, we can take a full backup of all resources in the namespace
+# Now that those backups are done, we can take a full backup of all resources in the namespace
 # In 001-barParameters.sh, one can specify which resources to skip, get that list first
 skipToBackupResourceKinds=$(echo $barSkipToBackupResourceKinds | tr "," "\n")
 logInfo "Collecting resources that need to be backed up..."
@@ -321,8 +387,7 @@ for i in $allResources; do
 done
 echo
 
-
-# Create the backup script header
+## Create the backup script header
 
 PV_BACKUP_DIR=${pvBackupDirectory}/$(basename $(dirname $BACKUP_DIR))/$(basename $BACKUP_DIR)
 
@@ -359,9 +424,80 @@ EOF
 oc get pvc -n $cp4baProjectName -o 'custom-columns=ns:.metadata.namespace,class:.spec.storageClassName,pv:.spec.volumeName,name:.metadata.name' --no-headers | sed 's/^/perform_backup /g' >> 023-backup-pvs.sh
 
 
+##### Scale down ###########################################
+printf "Do you want to scale down pods used during the backup? (Yes/No, default: No): "
+read -rp "" ans
+case "$ans" in
+"y"|"Y"|"yes"|"Yes"|"YES")
+   echo -e "\nScaling down pods used during backup in namespace ${cp4baProjectName}..."
+   ;;
+*)
+   echo -e "\nExiting...\n"
+   logInfo "All resources were backed up."
+   exit 0
+   ;;
+esac
+
+## Scale down CPfs
+if [[ $CP4BA_VERSION =~ "21.0.3" ]]; then
+  logInfo "Scaling down MongoDB pods..."
+  logInfo $(oc scale statefulset -l=app=icp-mongodb --replicas=0)
+  logInfo "Scaling down MetastoreDB pods..."
+  logInfo $(oc scale statefulset -l=component=zen-metastoredb --replicas=0);
+else # 24.0 and greater
+  #TODO: Scale down CPFS services in other versions of CP4BA.  
+  # Not aborting excecution for this though 
+  true 
+fi
+
+## BTS 
+logInfo "Scaling down last BTS PostgreSQL Database pod..."
+logInfo $(oc delete pod "ibm-bts-cnpg-"$cp4baProjectName"-cp4ba-bts-1")
+echo
+
+## Flink job submitters
+logInfo "Removing flink job submitters..."
+oc get jobs -o custom-columns=NAME:.metadata.name | grep bai- | grep -v bai-setup | xargs oc delete job
+echo
+  
+## Insights engine management 
+logInfo "Scaling down insights-engine pods..."
+logInfo $(oc scale deployment.apps/iaf-insights-engine-management --replicas=0);
+echo 
+
+## Kafka
+logInfo "Scaling down Kafka pods..."
+sts=$(oc get statefulset.apps/iaf-system-kafka -o name --ignore-not-found)
+if [[ "$sts" = "" ]]; then
+  logInfo $(oc delete pod iaf-system-kafka-0)
+else
+  logInfo $(oc scale statefulset.apps/iaf-system-kafka --replicas=0);
+fi
+echo
+
+## Zookeeper 
+logInfo "Scaling down Zookeeper pods..."
+sts=$(oc get statefulset.apps/iaf-system-zookeeper -o name --ignore-not-found)
+if [[ "$sts" = "" ]]; then
+  logInfo $(oc delete pod iaf-system-zookeeper-0)
+else
+  logInfo $(oc scale statefulset.apps/iaf-system-zookeeper --replicas=0);
+fi
+echo
+
+## Elasticsearch Or Opensearch 
+logInfo "Scaling down Elasticsearch/Opensearch pods..."
+if [[ $CP4BA_VERSION =~ "21.0.3" ]]; then
+  # Elasticsearch
+  logInfo $(oc scale statefulset.apps/iaf-system-elasticsearch-es-data --replicas=0);
+else # 24.0 and greater
+  #Opensearch
+  logInfo $(oc scale statefulset -l=app.kubernetes.io/name=opensearch --replicas=0);
+fi
+echo
+
 logInfo "All resources are backed up. Next, please back up:"
 logInfo "  - the databases"
 logInfo "  - the content of the PVs (by running the script 023-backup-pvs.sh on the storage server using the root account)"
 logInfo "  - the binary document data of CPE"
-logInfo "Once those backups are complete, pls. scale up the deployment again."
 echo
